refactor(keystore): rename `e2ei_acme_ca.rs` -> `x509_trust_anchor.rs`
